sensors/requirements.txt
click<8.1.0
paho-mqtt==1.6.1
typer==0.4.0
